requirements.txt
backcall==0.2.0
brotlipy==0.7.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
decorator==5.0.9
idna==2.10
ipykernel==5.3.4
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.17.0
jupyter-client==6.1.12
jupyter-core==4.7.1
olefile==0.46
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
pip==21.1.3
prompt-toolkit==3.0.17
pycparser==2.20
Pygments==2.9.0
pyOpenSSL==20.0.1
PySocks==1.7.1
python-dateutil==2.8.2
pywin32==227
pyzmq==20.0.0
requests==2.25.1
selenium==3.141.0
setuptools==52.0.0.post20210125
six==1.16.0
tornado==6.1
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
wheel==0.36.2
win-inet-pton==1.1.0
wincertstore==0.2
